% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_templates.R
\name{addHeaderR_header_template}
\alias{addHeaderR_header_template}
\title{Template helpers for addins}
\usage{
addHeaderR_header_template(level = 1)
}
\arguments{
\item{level}{Header level (1, 2, or 3).}
}
\description{
Template helpers for addins
}
\keyword{Header}
\keyword{internal}
\keyword{template}
