% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_templates.R
\name{addHeaderR_parse_numbered_header_line}
\alias{addHeaderR_parse_numbered_header_line}
\title{Parse a numbered RStudio section header line}
\usage{
addHeaderR_parse_numbered_header_line(line)
}
\arguments{
\item{line}{A single line.}
}
\value{
A list with elements \code{level} and \code{nums}, or NULL.
}
\description{
Parse a numbered RStudio section header line
}
\keyword{internal}
