% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_templates.R
\name{addHeaderR_roxygen_template}
\alias{addHeaderR_roxygen_template}
\title{Build a roxygen2 template}
\usage{
addHeaderR_roxygen_template(name, args)
}
\arguments{
\item{name}{Function name (may be empty).}

\item{args}{Character vector of argument names.}
}
\value{
A character string with roxygen lines.
}
\description{
Build a roxygen2 template
}
\keyword{internal}
