% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_generate_roxygen_from_selection.R
\name{generate_roxygen_from_selection}
\alias{generate_roxygen_from_selection}
\title{Generate roxygen2 template from selected function text}
\usage{
generate_roxygen_from_selection()
}
\value{
Invisibly, NULL.
}
\description{
Select an R function definition in the editor, then run this addin to insert a
roxygen2 template above it.
}
